Reduced the number of warnings and resolved some ToDos for ticket #601
git-svn-id: https://svn.code.sf.net/p/numere/source/branches/ticket_601@956 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/datamanagement/memory.hpp
@@ -85,8 +85,8 @@ class Memory : public Sorter
 		bool isValid() const;
 		bool isValidElement(size_t _nLine, size_t _nCol) const;
 		bool shrink();
-		size_t getLines(bool _bFull = false) const;
-		size_t getCols(bool _bFull = false) const;
+		int getLines(bool _bFull = false) const;
+		int getCols(bool _bFull = false) const;
         size_t getSize() const;
 
 
@@ -104,7 +104,7 @@ class Memory : public Sorter
 
 		// WRITE ACCESS METHODS
 		void writeSingletonData(Indices& _idx, const mu::value_type& _dData);
-		void writeData(size_t _nLine, size_t _nCol, const mu::value_type& _dData);
+		void writeData(int _nLine, int _nCol, const mu::value_type& _dData);
 		void writeData(Indices& _idx, mu::value_type* _dData, unsigned int _nNum);
 		bool setHeadLineElement(size_t _i, const std::string& _sHead);
